Move window to a file and clean up
